В бд теперь сохраняется только название фото

+++ src/main/java/com/concretepage/dao/PhotoDAO.java
@@ -1,5 +1,6 @@
 package com.concretepage.dao;
 
+import com.concretepage.controller.PhotoController;
 import com.concretepage.entity.Photo;
 import com.concretepage.entity.UserInfo;
 import org.springframework.stereotype.Repository;
@@ -7,6 +8,11 @@ import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import java.io.File;
+import java.io.IOException;
+import java.nio.file.DirectoryNotEmptyException;
+import java.nio.file.Files;
+import java.nio.file.NoSuchFileException;
 import java.util.List;
 
 @Transactional
@@ -15,11 +21,13 @@ public class PhotoDAO implements IPhotoDAO {
 
     @PersistenceContext
     private EntityManager entityManager;
+    private static String UPLOADED_FOLDER = "D://temp//";
 
     @Override
-    public List<Photo> getAllPhotos() {
-        String hql ="FROM Photo";
-        return (List<Photo>) entityManager.createQuery(hql).getResultList();
+    public List<Photo> getAllPhotos(int albumId) {
+        String hql ="FROM Photo as ph WHERE ph.albumId = ?";
+        return (List<Photo>) entityManager.createQuery(hql)
+                .setParameter(1, albumId).getResultList();
     }
 
     @Override
@@ -34,20 +42,32 @@ public class PhotoDAO implements IPhotoDAO {
 
     @Override
     public void updatePhoto(Photo photo) {
-        Photo ph = getPhotoById(photo.getPhoto_id());
-        /*ph.setLogin(photo.getLogin());
-        ph.setPassword(photo.getPassword());*/
+        Photo ph = getPhotoById(photo.getPhotoId());
         entityManager.flush();
     }
 
     @Override
     public void deletePhoto(int photoId) {
-        entityManager.remove(getPhotoById(photoId));
+        Photo photo = entityManager.find(Photo.class, photoId);
+        if(delete(photo)) {
+            entityManager.remove(getPhotoById(photoId));
+        }
+    }
+
+    private boolean delete(Photo photo){
+        File file = new File(UPLOADED_FOLDER + photo.getPhotoLink());
+        try {
+            return file.delete();
+        } catch (Exception ex) {
+            System.out.println(ex.getMessage());
+
+        }
+        return false;
     }
 
     @Override
     public boolean photoExists(String link) {
-        String hql = "FROM Photo as pht WHERE pht.photo_link = ?";
+        String hql = "FROM Photo as pht WHERE pht.photoLink = ?";
         int count = entityManager.createQuery(hql).setParameter(1, link)
                 .getResultList().size();
         return count > 0;
